GEODE-9214: Revision 1
 - Changed test/cpp logging to Boost.Log
 - Created modules to avoid including Boost.Log in headers from
 test/security.
 - Added security to old ITs to solve undefined symbols from
 test/security.
- Added C4596 warning supression for test/security that solves a
  inclusion issue related to Boost.Log in Windows platform.
