+++ tests/cpp/fwklib/FwkException.hpp
@@ -34,6 +34,14 @@
 
 // ----------------------------------------------------------------------------
 
+#define FWKEXCEPTION(x)                                              \
+  do {                                                               \
+    std::ostringstream os;                                           \
+    os << x << " In file: " << __FILE__ << " at line: " << __LINE__; \
+    throw apache::geode::client::testframework::FwkException(        \
+        os.str().c_str());                                           \
+  } while (0)
+
 namespace apache {
 namespace geode {
 namespace client {
