+++ tests/cpp/security/LdapUserCredentialGenerator.hpp
@@ -81,21 +81,10 @@ class LdapUserCredentialGenerator : public CredentialGenerator {
   std::string getClientDummyAuthorizer() override {
     return "javaobject.DummyAuthorization.create";
   }
-  void getValidCredentials(std::shared_ptr<Properties>& p) override {
-    p->insert("security-username", "geode1");
-    p->insert("security-password", "geode1");
-    FWKDEBUG("inserted valid security-username "
-             << p->find("security-username")->value().c_str() << " password "
-             << p->find("security-password")->value().c_str());
-  }
 
-  void getInvalidCredentials(std::shared_ptr<Properties>& p) override {
-    p->insert("security-username", "geode1");
-    p->insert("security-password", "1geode");
-    FWKDEBUG("inserted invalid security-username "
-             << p->find("security-username")->value().c_str() << " password "
-             << p->find("security-password")->value().c_str());
-  }
+  void getValidCredentials(std::shared_ptr<Properties>& p) override;
+
+  void getInvalidCredentials(std::shared_ptr<Properties>& p) override;
 
   void getAllowedCredentialsForOps(opCodeList& opCodes,
                                    std::shared_ptr<Properties>& p,
