+++ tests/cpp/security/XmlAuthzCredentialGenerator.hpp
@@ -174,112 +174,10 @@ class XmlAuthzCredentialGenerator {
   }
 
  private:
-  void getAllowedDummyAuthz(ROLES role) {
-    const char* adminUsers[] = {"admin", "root", "administrator"};
-    const int adminUsrSz = (sizeof adminUsers / sizeof *adminUsers) - 1;
-    std::string validity = "invalid";
-
-    if (role == NO_ROLE) {
-      role = getRequiredRole();
-      validity = "valid";
-    }
-    char userName[100];
-    switch (role) {
-      case READER_ROLE:
-        sprintf(userName, kPRiUsername, "reader", randomValue(2));
-        break;
-      case WRITER_ROLE:
-        sprintf(userName, kPRiUsername, "writer", randomValue(2));
-        break;
-      case QUERY_ROLE:
-        sprintf(userName, kPRiUsername, "reader", randomValue(1) + 3);
-        break;
-      case ADMIN_ROLE:
-        sprintf(userName, "%s", adminUsers[randomValue(adminUsrSz)]);
-        break;
-      case NO_ROLE:
-        sprintf(userName, kPRiUsername, "user", randomValue(2));
-        break;
-    }
-
-    (*m_prop)->insert("security-username", userName);
-    (*m_prop)->insert("security-password", userName);
-
-    FWKINFO("inserted "
-            << validity << " dummy security-username "
-            << (*m_prop)->find("security-username")->value().c_str()
-            << " password "
-            << ((*m_prop)->find("security-password") != nullptr
-                    ? (*m_prop)->find("security-password")->value().c_str()
-                    : "not set"));
-  }
-
-  std::string getAllowedUser(ROLES role) {
-    const std::string userPrefix = "geode";
-    const int readerIndices[] = {3, 4, 5};
-    const int writerIndices[] = {6, 7, 8};
-    const int queryIndices[] = {9, 10};
-    const int adminIndices[] = {1, 2};
-    const int readerIndSz = (sizeof readerIndices / sizeof *readerIndices) - 1;
-    const int writerIndSz = (sizeof writerIndices / sizeof *writerIndices) - 1;
-    const int queryIndSz = (sizeof queryIndices / sizeof *queryIndices) - 1;
-    const int adminIndSz = (sizeof adminIndices / sizeof *adminIndices) - 1;
-
-    std::string validity = "invalid";
-
-    if (role == NO_ROLE) {
-      role = getRequiredRole();
-      validity = "valid";
-    }
-    char userName[256];
-    switch (role) {
-      case READER_ROLE:
-        sprintf(userName, kPRiUsername, userPrefix.c_str(),
-                readerIndices[randomValue(readerIndSz)]);
-        break;
-      case WRITER_ROLE:
-        sprintf(userName, kPRiUsername, userPrefix.c_str(),
-                writerIndices[randomValue(writerIndSz)]);
-        break;
-      case QUERY_ROLE:
-        sprintf(userName, kPRiUsername, userPrefix.c_str(),
-                queryIndices[randomValue(queryIndSz)]);
-        break;
-      case ADMIN_ROLE:
-      case NO_ROLE:
-        sprintf(userName, kPRiUsername, userPrefix.c_str(),
-                adminIndices[randomValue(adminIndSz)]);
-        break;
-    }
-    FWKINFO("inserted " << validity << " username " << userName);
-    return std::string(userName);
-  }
-
-  void getAllowedLdapAuthz(ROLES role) {
-    const std::string userName = getAllowedUser(role);
-    (*m_prop)->insert("security-username", userName.c_str());
-    (*m_prop)->insert("security-password", userName.c_str());
-
-    FWKINFO("inserted  ldap security-username "
-            << (*m_prop)->find("security-username")->value().c_str()
-            << " password "
-            << ((*m_prop)->find("security-password") != nullptr
-                    ? (*m_prop)->find("security-password")->value().c_str()
-                    : "not set"));
-  }
-
-  void getAllowedPkcsAuthz(ROLES role) {
-    const std::string userName = getAllowedUser(role);
-    (*m_prop)->insert("security-alias", userName.c_str());
-    (*m_prop)->insert("security-keystorepass", "geode");
-
-    FWKINFO("inserted  PKCS security-alias"
-            << (*m_prop)->find("security-alias")->value().c_str()
-            << " password "
-            << ((*m_prop)->find("security-keystorepass") != nullptr
-                    ? (*m_prop)->find("security-keystorepass")->value().c_str()
-                    : "not set"));
-  }
+  void getAllowedDummyAuthz(ROLES role);
+  std::string getAllowedUser(ROLES role);
+  void getAllowedLdapAuthz(ROLES role);
+  void getAllowedPkcsAuthz(ROLES role);
 
   ROLES getRequiredRole() {
     bool requireReaders = true, requireWriters = true, requireQuery = true;
